Refactor + styling
