+++ src/common/components/core/ChatWidget/ChatWidget.tsx
@@ -1,12 +1,14 @@
 import { useEffect, useState } from 'react';
 import { addResponseMessage, addLinkSnippet, Widget } from 'react-chat-widget';
-import { useAskChatGuidanceQuestionQuery } from '../../../core/apollo/generated/apollo-hooks';
+import { useAskChatGuidanceQuestionQuery } from '../../../../core/apollo/generated/apollo-hooks';
 
 import logo from './favicon-16x16.png';
 import { useTranslation } from 'react-i18next';
-import { FEATURE_GUIDANCE_ENGINE } from '../../../domain/platform/config/features.constants';
-import { useConfig } from '../../../domain/platform/config/useConfig';
-import { useUserContext } from '../../../domain/community/contributor/user';
+import { FEATURE_GUIDANCE_ENGINE } from '../../../../domain/platform/config/features.constants';
+import { useConfig } from '../../../../domain/platform/config/useConfig';
+import { useUserContext } from '../../../../domain/community/contributor/user';
+import 'react-chat-widget/lib/styles.css';
+import './styles.css';
 
 const ChatWidget = () => {
   const [newMessage, setNewMessage] = useState(null);
