+++ src/common/components/core/ChatWidget/ChatWidgetQueries.graphql
